backend/requirements.txt
blinker==1.9.0
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3

psycopg2-binary==2.9.9

